기존 store 삭제, charts system section add 기능 필요
