+++ src/components/WorkSpace/HPOPage.js
@@ -1,8 +1,7 @@
 import React, { useEffect } from 'react';
 import RunTableComponent from './RunTableComponent';
 import Header from './WorkSpaceHeader';
-import SectionsComponent from './sections/SectionsComponent';
-import SystemComponent from './system/SystemComponent';
+
 import { useDispatch, useSelector } from 'react-redux';
 
 import * as Actions from '../../data/model/actions.js';
@@ -11,7 +10,7 @@ import * as Selectors from '../../data/model/selectors.js';
 import { useParams } from 'react-router-dom';
 
 const HPOListComponents = () => {
-  const { id } = useParams();
+  const { id } = useParams(); 
   console.log(id);
 
   const dispatch = useDispatch();
