+++ src/components/WorkSpace/system/SystemComponent.js
@@ -1,22 +1,20 @@
 import React from 'react';
+import SystemCards from './SystemCards';
 
-import SystemComponents from './SystemComponents';
-import { useSelector } from 'react-redux';
 
 const SystemComponent = ({ models, graph, isLoading, cards }) => {
-  const systems = useSelector((state) => state.system.cards);
+ 
 
   return (
     <div style={{ paddingBottom: '3rem' }}>
-      {systems.map((system, index) => (
-        <SystemComponents
-          key={index}
+      
+        <SystemCards
           models={models}
           graph={graph}
           isLoading={isLoading}
           cards={cards}
-        ></SystemComponents>
-      ))}
+        ></SystemCards>
+    
     </div>
   );
 };
