--- src/store/actions/Section.action.js
@@ -1,6 +1,0 @@
-export const FETCH_CARD_DATA = "sectionCard/FETCH_CARD_DATA";
-
-export const fectchCardData = (card) => {
-  return { type: FETCH_CARD_DATA, card };
-};
-
