--- src/store/actions/System.action.js
@@ -1,5 +1,0 @@
-export const SYSTEM_CARD = "systemcard/SYSTEM_CARD";
-
-export const systemCard = (card) => {
-  return { type: SYSTEM_CARD, card };
-};
