--- src/store/actions/index.js
@@ -1,5 +1,0 @@
-export * from "./Section.action";
-
-export * from "./System.action";
-export * from "./Model.action";
-export * from "./Project.action";
